requirements.txt
awscli==1.16.291
boto==2.49.0
botocore==1.13.27
tqdm==4.36.1
pillow==6.2.1